djmukmenus/requirements.txt
# Python Package Requirements File
# More info: http://pypi.python.org/pypi/pip

# From the PyPi repository 
# Example: ``pip install Django``

django-debug-toolbar
django-endless-pagination
django-extensions
django-tagging
django-tagging
Django==1.3.1
Fabric
gunicorn
MySQL-python
PIL
python-dateutil==1.5 # higher versions are not compat with py 2.x 
raven>=1.4.6 # Dep: Sentry Logging Service
simplejson
sorl-thumbnail
South
